# Add in-place multiply-add