This adds support for the new 5-argument in-place matrix multiplication in Julia v1.3 (#357 ).
We can probably consolidate the multiply-add and with the matrix multiply methods, but I've left them separate for now.

I've also added a quick "hack" to avoid an invalid pointer error (#734) when doing in-place multiplication with SizedArrays large enough that they get passed to BLAS.