@andyferris 2020-02-19
Certainly seems like a pretty good idea - did you do any benchmarking comparing performance to naive multiply-then-add?