@bjack205 2020-02-19
Yep, here's a few benchmarks

Test script:
```julia
julia> α,β = 2.0, 1.0
julia> N = 3 
julia> a = @MMatrix rand(N,N)
julia> b = @MMatrix rand(N,N)
julia> c = @MMatrix rand(N,N)
julia> A,B,C = Array(a), Array(b), Array(c)
julia> @btime mul!($C, $A, $B, 1.0, 1.0);
julia> @btime mul!($c, $a, $b, 1.0, 1.0);
```
And the results:
### N = 3
| Version | @btime result |
| ---------- | ------------------ |
| Base Array | 16.470 ns (0 allocations: 0 bytes) |
|  Master branch | 11.839 ns (0 allocations: 0 bytes) |
| This version | 8.847 ns (0 allocations: 0 bytes) |

### N = 5
| Version | @btime result |
| ---------- | ------------------ |
| Base Array | 121.191 ns (0 allocations: 0 bytes) |
| Master branch | 182.495 ns (4 allocations: 224 bytes) |
| This version | 22.558 ns (0 allocations: 0 bytes) |

### N = 20
| Version | @btime result |
| ---------- | ------------------ |
| Base Array | 581.967 ns (0 allocations: 0 bytes) |
| Master branch | 4.618 μs (4 allocations: 224 bytes) |
| This version | 575.928 ns (0 allocations: 0 bytes) |