@bjack205 2020-03-06
Thanks for the feedback! This is my first PR so I certainly have a lot to learn!

I've rehashed my solution and made it shorter and hopefully more along the lines of what you're looking for. I've merged the 5-argument `mul!` with the 3-argument `mul!` and unified the way transposes are handled. 

The transposes are particularly tricky since they tend to allocate at the slightest provocation. My solution avoids this allocation with a function barrier and instead caches a transpose flag in with the type (see `TSize` in `matrix_multiply_add.jl`). This ended up with a fairly straight-forward solution, IMHO.

The benchmarks are pretty much the same as they were before. 
The three-argument multiplication doesn't incur any performance hit for small matrices, but for the medium-sized "chunked" version, my implementation runs a little slower.