@bjack205 2020-03-06
I missed a test case that happens to be breaking pretty much everything. All my tests passed in magic numbers for alpha and beta, but when they're variables the `MulAddMul` allocates (due to type instability). I've tried a few things but can't figure out how to recover from the type instability. I'm not familiar enough with Julia to really understand why my implementation doesn't work while the one in `LinearAlgebra/matmul.jl` works, since they seem fairly similar to me. I'm open to any ideas on how to fix this, because otherwise this attempt isn't worth much.