@bjack205 2020-03-07
Okay, I think it should be pretty close with the most recent commit. I hard-coded the `MulAddMul` for both the 5-argument and 3-argument versions to avoid type instability and resultant allocations. 

NOTE THAT USING `MulAddMul` IS A BREAKING CHANGE, since this is a feature in Julia v1.3. I don't see this as a problem, since the 5-argument `mul!` was just made available in v1.3, anyways.

I also fixed the performance decline for the chunked unrolling, so the new version has improved or equivalent performance on all sizes. 

I guess I should also give some motivation for why I think these methods are important. In my package `TrajectoryOptimization.jl` I've been battling getting the core algorithm to not allocate any memory. Normal `SArrays` work great, but we found that compilation times went through the roof when we tried larger problems. To my knowledge (which, admittedly, isn't all that much when it comes to Julia), `SizedArray` with `mul!` is the only solution that works nearly as fast as `SArrays` at small sizes (thanks to loop unrolling!), scales well to large arrays (thanks to BLAS), and doesn't incur any memory allocations.

Newest benchmarks:
test script:
```julia
function benchmark_matmul(N1,N2,ArrayType=MArray)
    if ArrayType <: MArray
        Mat = MMatrix
        Vec = MVector
    elseif ArrayType <: SizedArray
        Mat = SizedMatrix
        Vec = SizedVector
    end
    α,β = 1.0, 1.0
    A = rand(Mat{N1,N2})
    B = rand(Mat{N2,N2})
    C = rand(Mat{N1,N2})
    println("C = A*B")
    @btime mul!($C,$A,$B)
    println("C += A*B")
    @btime mul!($C,$A,$B,$α,$β)
    println("B = A'C")
    @btime mul!($B,Transpose($A),$C)
    println("B += A'C")
    @btime mul!($B,Transpose($A),$C,$α,$β)
end

```
### `MArray`
| Size | branch | `C = A*B` | `C += A*B` | `B = A'B` | `B += A'B` |
| ---- | ---- | ---- | ---- | --- | ---- | 
| 3x3 | master | 6.6 ns (0 allocs) | 193.4 ns (3 allocs) | 9.2 ns (0 allocs) | 193.8 ns (3 allocs) |
| 3x3 | new      | 6.7 ns (0 allocs) | 8.1 ns (0 allocs)    | 6.6 ns (0 allocs) | 8.1 ns (0 allocs) |
| 10x10 | master | 123.3 ns (0 allocs) | 889.5 ns (7 allocs) | 489.6 ns (0 allocs) | 683.9 ns (3 allocs) |
| 10x10 | new      | 124.1 ns (0 allocs) | 134.1 ns (0 allocs) | 196.9 ns (0 allocs) | 207.2 ns (3 allocs) |
| 20x20 | master | 613.7 ns (0 allocs) | 4.83 μs (7 allocs) | 3.76 μs (0 allocs) | 4.21 μs (3 allocs) |
| 20x20 | new      | 613.7 ns (0 allocs) | 580.9 ns (0 allocs) | 613.4 ns (0 allocs) | 591 ns (3 allocs) |

### `SizedArray`
| Size | branch | `C = A*B` | `C += A*B` | `B = A'B` | `B += A'B` |
| ---- | ---- | ---- | ---- | --- | ---- | 
| 3x3 | master | 10 ns (0 allocs) | 195 ns (3 allocs) | 10 ns (0 allocs) | 194 ns (0 allocs) |
| 3x3 | new | 10 ns (0 allocs) | 11 ns (0 allocs) | 11 ns (0 allocs) | 11 ns (0 allocs) |
| 10x10 | master | 119 (0 allocs) | 905 ns (7 allocs) | 406 ns (0 allocs) | 680 ns (3 allocs) |
| 10x10 | new | 125 ns (0 allocs) | 129 ns (0 allocs) | 195 ns (0 allocs) | 199 ns (0 allocs) |
| 20x20 | master | error | error | error | error |
| 20x20 | new | 617 ns (0 allocs) | 581 ns (0 allocs) | 612 ns (0 allocs) | 576 ns (0 allocs) |