@c42f 2020-04-14
Oops it looks like I caused a conflict by merging some unrelated test tools. I merged into your branch which will hopefully fix that up.

Generally this is looking quite good now (a lot less scary in terms of amount of code added!)

Great job with the table of benchmarks there's some really impressive improvements there. Could you add the benchmarks to `benchmark` directory so that we retain a copy of them and run them in the future? You can see the results of comparing benchmarks across versions in the github actions "Benchmark" job. There's other examples there of how to hook them up.