@c42f 2020-04-17
I went ahead and merged this, we can sort the `@allocated` mysteries out another time.

Thanks for a very nice contribution!